+++ photorealism-plugin/tools/shader_check.sh
@@ -50,6 +50,7 @@ shaders=(
   "ssao.hlsl PSSSAO ps_5_0"
   "temporal.hlsl PSTemporal ps_5_0"
   "rtgi.hlsl PSRtgi ps_5_0"
+  "rtgi.hlsl PSRtgiCompose ps_5_0"
 )
 
 cd "${project_dir}/shaders"